Refactor(Testes): Ajustando as pastas dos testes e corrigindo erros nas classes dos testes.
